requirements.txt
amqp==5.0.9
asgiref==3.2.10
attrs==21.4.0
autobahn==22.1.1
Automat==20.2.0
backports.zoneinfo==0.2.1
billiard==3.6.4.0
blinker==1.4
boto3==1.20.46
botocore==1.23.46
Brlapi==0.7.0
celery==5.2.3
certifi==2019.11.28
cffi==1.15.0
channels==3.0.4
chardet==3.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
constantly==15.1.0
cryptography==36.0.1
cycler==0.11.0
daphne==3.0.2
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.4
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1
django-allauth==0.46.0
django-cors-headers==3.11.0
django-countries==7.2.1
django-filter==21.1
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-passwordreset==1.2.1
django-shortuuidfield==0.1.3
django-test-plus==2.2.0
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
djangorestframework-simplejwt==5.1.0
dynaconf==3.1.7
entrypoints==0.3
filelock==3.4.2
fonttools==4.29.0
gpg===1.13.1-unknown
httplib2==0.14.0
hvac==0.11.2
hyperlink==21.0.0
idna==2.8
incremental==21.3.0
inflection==0.5.1
install==1.3.5
Jinja2==3.0.3
jmespath==0.10.0
keyring==18.0.1
kiwisolver==1.3.2
knox==0.1.14
kombu==5.2.3
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
loguru==0.6.0
louis==3.12.0
MarkupSafe==2.0.1
matplotlib==3.5.0
mysqlclient==2.1.0
netifaces==0.10.4
numpy==1.22.1
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
opencv-python==4.5.5.62
packaging==21.3
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.4.1
prompt-toolkit==3.0.27
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
PyGObject==3.36.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-apt===2.0.0-ubuntu0.20.04.6-elementary6-ubuntu6.1
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2021.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
redis==4.1.3
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
s3transfer==0.5.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
service-identity==21.1.0
setuptools-scm==6.4.2
shortuuid==1.0.8
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.2
tomli==2.0.0
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
ubuntu-advantage-tools==27.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
validators==0.18.2
vine==5.0.0
virtualenv==20.13.0
wadllib==1.3.3
wcwidth==0.2.5
wrapt==1.13.3
xkit==0.0.0
youtube-dl==2020.3.24
zope.interface==5.4.0
